backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
django-debug-toolbar==3.8.1
